requirements.txt
fastapi
uvicorn
sqlalchemy  
python-dotenv
alembic
pgvector
psycopg2